Fix #16 Accept command-line arguments.
If main.php is called with commandline arguments, these will be passed
to the bridge as a single request, instead of running the bridge as a daemon.

The arguments are, in order: <command> <username> <server> <password>

The result will be printed to stdout (1 for success, 0 for failure),
and will be set as the exit status (0 for success, 1 for failure).